+++ examples/unix_socket.js
@@ -1,5 +1,6 @@
 var redis = require("redis"),
-    client = redis.createClient("/tmp/redis.sock");
+    client = redis.createClient("/tmp/redis.sock"),
+    profiler = require("v8-profiler");
 
 client.on("connect", function () {
     console.log("Got Unix socket connection.")
@@ -9,7 +10,20 @@ client.on("error", function (err) {
     console.log(err.message);
 });
 
-client.info(function (err, reply) {
-    console.log(reply.toString());
-    client.quit();
-});
+client.set("space chars", "space value");
+
+setInterval(function () {
+    client.get("space chars");
+}, 100);
+
+function done() {
+    client.info(function (err, reply) {
+        console.log(reply.toString());
+        client.quit();
+    });
+}
+
+setTimeout(function () {
+    console.log("Taking snapshot.");
+    var snap = profiler.takeSnapshot();
+}, 5000);